backend/requirements.txt
flask==3.1.1
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.0
flask-login==0.6.3
flask-bcrypt==1.0.1 